More optimization and batch rendering
Removed redundant translations to the MVP matrix. Now the chunks models and the custom models are translated before being sent to the GPU and the MVP matrix is now updated only once per frame.

Batch rendering is now in alpha. It renders more than one model per batch but it is necessary to add a method to update the models inside a batch.

Also, block selection is broken due to an error in the ray casting algorithm implementation. Will try to fix this in the following update.