GUI
Started adding support for both 2D and 3D rendering. This obviously has a performance cost since I have to keep track of when I have to render with the projection matrix and when with the ortho matrix. But hey, it's a start towards having a GUI and a save/load world button and a selector of blocks :D